tests/wpt/tests/css/css-text/parsing/text-align-last-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text: getComputedStyle().textAlignLast</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-text-align-last">
<meta name="assert" content="text-align-last computed value is specified keyword.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("text-align-last", "auto");
test_computed_value("text-align-last", "start");
test_computed_value("text-align-last", "end");
test_computed_value("text-align-last", "left");
test_computed_value("text-align-last", "right");
test_computed_value("text-align-last", "center");
test_computed_value("text-align-last", "justify");
test_computed_value("text-align-last", "match-parent");
</script>
</body>
</html>
